refactor(store-core): split misc.ts handler catch-all into topical slices (#6034)
The P2-3 family split (#5892-#5913) left handlers/misc.ts as an
unnamed catch-all — new handlers defaulted into it, re-accreting the exact
anti-pattern the split removed elsewhere. Split it into four cohesively-named
slices, each following the existing family-file pattern (self-contained
imports, re-exported via the ./index barrel):

- session-config.ts — session/agent runtime config + readiness:
  handleModelChanged, handleClaudeReady, handleAgentIdle, handleAgentBusy,
  handleThinkingLevelChanged (+ ThinkingLevel).
- inventory.ts — server-side inventory list-replacement:
  handleSlashCommands, handleAgentList, handleProviderList, handleFileList,
  handleAvailableModels (+ AvailableModelsPayload), handleMcpServers, plus the
  private shouldSkipForSessionMismatch guard.
- alerts.ts — client-side alert / banner state:
  handleSessionCostThresholdCrossed (+ SessionCostThresholdCrossedPayload),
  handleNotificationPrefs (+ NotificationPrefsState, NotificationPrefsPayload).
- permission-stream-split.ts — the #554 resolvePermissionStreamSplit
  mid-stream permission-boundary split resolver.

Pure file move: function bodies, comments, and types are copied verbatim. The
set of exported symbols is byte-for-byte identical to the old misc.ts, so the
./index barrel's public surface (the single import surface tests and both
clients use) is unchanged. No dispatch wiring changes — consumers import from
the barrel, never from misc directly.

Verified: tsc --noEmit clean; full store-core vitest suite green
(40 files, 1645 tests).
