+++ packages/store-core/src/handlers/session-config.ts
@@ -1,0 +1,116 @@
+/**
+ * Shared stateless handlers for session/agent runtime-config + readiness
+ * messages (model_changed, claude_ready, agent_idle / agent_busy,
+ * thinking_level_changed).
+ *
+ * Extracted from ./misc.ts (issue #6034 — splitting the P2-3 leftover
+ * catch-all into cohesively-named slices). Pure move, no logic change.
+ * Re-exported from ./index so the public surface is unchanged. See ./index.ts
+ * for the stateless-handler contract.
+ */
+
+import type { ActiveTool, TranscriptBackgroundTask } from '../types'
+import { parseStringField } from './_shared'
+
+// ---------------------------------------------------------------------------
+// model_changed
+// ---------------------------------------------------------------------------
+
+/** Extract the model value from a `model_changed` message. */
+export function handleModelChanged(msg: Record<string, unknown>): { model: string | null } {
+  return { model: parseStringField(msg, 'model') }
+}
+
+// ---------------------------------------------------------------------------
+// claude_ready
+// ---------------------------------------------------------------------------
+
+/**
+ * State patch for `claude_ready`.
+ *
+ * `stoppedAt`/`stoppedCode` are reset to null here so the quiet "Session
+ * stopped." status strip introduced for #4879 clears the moment the
+ * server reports the child is ready again (typically because the user
+ * sent another message after tapping Stop). This is purely additive for
+ * sessions that were never stopped — both fields stay null end-to-end.
+ */
+export function handleClaudeReady(msg?: Record<string, unknown>): {
+  claudeReady: true
+  stoppedAt: null
+  stoppedCode: null
+  transcriptBackgroundTasks?: TranscriptBackgroundTask[]
+  scheduledWakeup?: { at: number; reason: string } | null
+} {
+  const patch: ReturnType<typeof handleClaudeReady> = {
+    claudeReady: true,
+    stoppedAt: null,
+    stoppedCode: null,
+  }
+  // #5431: enriched ready — `backgroundTasks` present (even as []) means the
+  // server computed a fresh transcript snapshot, so it is authoritative for
+  // BOTH fields: a snapshot with tasks but no wakeup means any previously
+  // stored wakeup has fired/been superseded. Absent means a pre-#5431 server
+  // or no transcript access — leave the stored fields untouched.
+  if (Array.isArray(msg?.backgroundTasks)) {
+    patch.transcriptBackgroundTasks = msg.backgroundTasks.filter(
+      (t): t is TranscriptBackgroundTask =>
+        !!t && typeof t === 'object' &&
+        typeof (t as TranscriptBackgroundTask).toolUseId === 'string' &&
+        ['bash', 'agent', 'monitor'].includes((t as TranscriptBackgroundTask).kind) &&
+        typeof (t as TranscriptBackgroundTask).description === 'string' &&
+        Number.isFinite((t as TranscriptBackgroundTask).startedAt),
+    )
+    const wakeup = msg.scheduledWakeup as { at?: unknown; reason?: unknown } | undefined
+    patch.scheduledWakeup =
+      wakeup && typeof wakeup.at === 'number' && typeof wakeup.reason === 'string'
+        ? { at: wakeup.at, reason: wakeup.reason }
+        : null
+  }
+  return patch
+}
+
+// ---------------------------------------------------------------------------
+// agent_idle / agent_busy
+// ---------------------------------------------------------------------------
+
+/** State patch for `agent_idle`.
+ *
+ * Also clears `streamingMessageId` so the stop button hides if the agent
+ * reaches idle without a closing `stream_end`/`result` (abnormal Agent SDK
+ * shutdown). Pre-#3170 the 5s safety timer in `sendInput` recovered this
+ * case; post-#3170 the timer is bypassed once `tool_start` bumps the value,
+ * so `agent_idle` is the remaining recovery hook. See #3171.
+ *
+ * #4308 — also clears `activeTools` as a safety net: a missed `tool_result`
+ * (server crash mid-turn, dropped broadcast, etc.) would otherwise leave a
+ * phantom "Running X" indicator visible for the rest of the session. Idle
+ * is a guaranteed turn-boundary, so it's the right place to drop any
+ * still-tracked in-flight tools.
+ */
+export function handleAgentIdle(): {
+  isIdle: true
+  streamingMessageId: null
+  activeTools: ActiveTool[]
+} {
+  return { isIdle: true, streamingMessageId: null, activeTools: [] }
+}
+
+/** State patch for `agent_busy`. */
+export function handleAgentBusy(): { isIdle: false } {
+  return { isIdle: false }
+}
+
+// ---------------------------------------------------------------------------
+// thinking_level_changed
+// ---------------------------------------------------------------------------
+
+export type ThinkingLevel = 'default' | 'high' | 'max'
+
+const VALID_THINKING_LEVELS = new Set<ThinkingLevel>(['default', 'high', 'max'])
+
+/** Extract and validate the thinking level from a `thinking_level_changed` message. */
+export function handleThinkingLevelChanged(msg: Record<string, unknown>): { level: ThinkingLevel } {
+  const raw = parseStringField(msg, 'level') || 'default'
+  const level = VALID_THINKING_LEVELS.has(raw as ThinkingLevel) ? (raw as ThinkingLevel) : 'default'
+  return { level }
+}
